change route name and fixed bug
